Delete exception

+++ src/main/java/com/roger/workshopspringbootjpa/services/UserService.java
@@ -2,8 +2,11 @@ package com.roger.workshopspringbootjpa.services;
 
 import com.roger.workshopspringbootjpa.entities.User;
 import com.roger.workshopspringbootjpa.repositories.UserRepository;
+import com.roger.workshopspringbootjpa.services.exceptions.DatabaseException;
 import com.roger.workshopspringbootjpa.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +32,15 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        userRepository.deleteById(id);
+        try {
+            if (userRepository.existsById(id)) {
+                userRepository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User user){
